refactor: modularize frontend components and backend command handlers, and add anyhow for improved error handling
