+++ wazuh-agent-status-rust/src-tauri/Cargo.toml
@@ -23,5 +23,6 @@ tauri-plugin-opener = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 image = "0.25"
+anyhow = "1.0"
 tauri-plugin-log = "2.8.0"
 tauri-plugin-positioner = { version = "2.3.1", features = ["tray-icon"] }
